Avoid looking for protobuf, kafka, grpc if requested
This fixes the issue some people are having when building monitoring without
online components (e.g. for analysis), while having protobuf / grpc / abseil on the system
for whatever reason.

In that scenario, the old behavior results in picking up a potentially incompatible
protobuf from the system and consequent build errors.

In order to keep backward compatible behavior, I made the options on by default,
I will make sure in alidist the correct options are chosen when building a non-dataflow
configuration.